# api_doc.yml
openapi: 3.0.1
info:
  title: Data Logger Server
  description: 'This is a central server for the use of a data storage and data processing.'
  contact:
    email: maximilian.nessen@uni-jena.de
  version: 1.0.0
servers:
- url: https://api.datalogger.de
...
paths:
  /handshake/hello:
   get:
      tags:
      - handshake
      summary: Does the Handshake for a new client
      description: Adds a new device to the system
      operationId: add Device
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Hello'
        500:
          description: Server error
          content: {}
      security:
      - petstore_auth:
        - write:pets
        - read:pets
  /handshake/register:
    post:
      tags:
      - handshake
      summary: Register a new client
      operationId: registerStation
      requestBody:
        description: Public Key of Station and new Id
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Register'
        required: true
      responses:
        200:
          description: Client register success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        405:
          description: Invalid input
          content: {}
        500: 
          description: Server error
          content: {}
      security:
      - petstore_auth:
        - write:pets
        - read:pets
...
components:
  schemas:
    Hello:
      type: object
      properties:
        pubic: 
         type: string
         description: Public RSA Key Server
         example: "-----BEGIN PUBLIC KEY-----
                    MIGfMA0GCSqGSIb3DQEBAQUAA4GNADCBiQKBgQCqGKukO1De7zhZj6+H0qtjTkVxwTCpvKe4eCZ0
                    FPqri0cb2JZfXJ/DgYSF6vUpwmJG8wVQZKjeGcjDOL5UlsuusFncCzWBQ7RKNUSesmQRMSGkVb1/
                    3j+skZ6UtW+5u09lHNsj6tQ51s1SPrCBkedbNf0Tp0GbMJDyR4e9T04ZZwIDAQAB
                    -----END PUBLIC KEY-----"
    Register:
      required:
      - public
      - station_id
      type: object
      properties:
        public: 
          description: Public RSA Key Client
          type: string
          example: "-----BEGIN PUBLIC KEY-----
                    MIGfMA0GCSqGSIb3DQEBAQUAA4GNADCBiQKBgQCqGKukO1De7zhZj6+H0qtjTkVxwTCpvKe4eCZ0
                    FPqri0cb2JZfXJ/DgYSF6vUpwmJG8wVQZKjeGcjDOL5UlsuusFncCzWBQ7RKNUSesmQRMSGkVb1/
                    3j+skZ6UtW+5u09lHNsj6tQ51s1SPrCBkedbNf0Tp0GbMJDyR4e9T04ZZwIDAQAB
                    -----END PUBLIC KEY-----"
        station_id:
          description: MAC Adress of Client
          type: string
          example: "00-14-22-01-23-45."
    RegisterResponse:
     type: object
     properties:
      secret:
        type: string
        description: Secret for encryption and decryption
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: http://petstore.swagger.io/oauth/dialog
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header